Add Kubernetes configuration files
simple.yaml is a Kubernetes configuration for just the worker -
it expects the DB host IP address to be provided in the config
file, and only the worker will be run in Kubernetes.

fully.yaml is a Kubernetes configuration for all of CoreKube -
the worker, frontend and DB are orchestrated in Kubernetes.
Currently this version is non-functional due to issues with the
way that Kubernetes handles SCTP traffic.