Fix multithreading bug involving statically-allocated buffers
Fix a bug where the raw contents of a message could be replaced
whilst a thread was processing them. This was due to buffer being
defined as a statically-allocated variable in start_listener()
which was then passed to all of the threads.

The solution is to dynamically allocate memory for a buffer for
each thread.