Add locks when accessing DB from multiple threads
With the introduction of multithreading, it is possible for multiple
threads to be accessing the DB at the same time. However, since the
DB is accessed using a single UDP connection, regardless of which
threads are accessing it, the DB accesses must therefore be limited
so that only one can occur at any given time.