Fix memory leak due to multithreaded changes
The changes made to fix a multithreaded bug in the previous commit
lead to a memory leak. There are several dynamically-allocated
variables that are freed at the end of the process_message()
function. However, if the message being processed was of type
NO_RESPONSE then (prior to this fix) the execution would return
from process_message() early, preventing the variables from being
freed, leading to the memory leak.