Ensure logging of incoming messages does not break with threading
Prior to this change, the printing / logging of the parsed S1AP
messages was completed using raw writes to stdout. The problem with
this approach is that if another thread writes to stdout during
this output, then the two outputs will become interleaved.

THis change prevents any other writes to stdout until the logging
of the incoming parsed S1AP message has been completed.