Fixed bug with DB logging in nas_attach.c
The DB logging had been put in an incorrect place in the file
nas_attach.c.

There are two DB accesses in nas_attach.c. The logging had been
placed around the access that only stored values in the DB, and
there was no logging around the accesses that retreived values
from the DB.

This is incorrect - there should be no logging around a DB store
(because the log analysis assumes that all logs are a retreive,
so the store will result in lower-than-usual timings polluting
the analysis), and there should be logging around the DB retreive.

This has now been corrected.