Bugfix: wrong DB values returned by Handover Required
There was a subtle yet serious bug in the Handover Required DB
code, due to the way that the Handover Required has two DB accesses:
a pull immediately followed by a pull. (This is necessary because
the eNB socket is retreived from the eNB DB and then immediately
stored into the UE DB).

Prior to this bugfix, the same buffer was used for both DB accesses.
This caused a bug because the extract_db_values() function returns
a struct containing many pointers to the original buffer (e.g: a
pointer to the KNH in the buffer, a pointer to the KASME in the
buffer, etc...). Therefore, reusing the same buffer after retreiving
those values would cause the buffer to be partially overwritten,
meaning that when the pointers to the retreived values were
accessed, they would return garbage data.

Correcting this bug was especially difficult because the second DB
access only overwrote the first few items in the buffer, and these
items happened to correspond to security keys (KNH, KASME) that
are random-looking values, so it was not obvious that they contained
garbage data.